+++ ClrHeapAllocationsAnalyzer.Test/DisplayClassAllocationAnalyzerTests.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Immutable;
+using ClrHeapAllocationAnalyzer.Analyzers;
 
 namespace ClrHeapAllocationAnalyzer.Test
 {
