+++ ClrHeapAllocationsAnalyzer.Test/TestConfiguration.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using ClrHeapAllocationAnalyzer.Support;
 
 namespace ClrHeapAllocationAnalyzer.Test
 {
