+++ ClrHeapAllocationsAnalyzer/Analyzers/EnumeratorAllocationAnalyzer.cs
@@ -1,13 +1,13 @@
-﻿namespace ClrHeapAllocationAnalyzer
-{
-    using System;
-    using System.Collections.Immutable;
-    using System.Linq;
-    using Microsoft.CodeAnalysis;
-    using Microsoft.CodeAnalysis.CSharp;
-    using Microsoft.CodeAnalysis.CSharp.Syntax;
-    using Microsoft.CodeAnalysis.Diagnostics;
+﻿using System;
+using System.Collections.Immutable;
+using System.Linq;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Diagnostics;
 
+namespace ClrHeapAllocationAnalyzer.Analyzers
+{
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
     public sealed class EnumeratorAllocationAnalyzer : AllocationAnalyzer
     {
