+++ ClrHeapAllocationsAnalyzer/Analyzers/ExplicitAllocationAnalyzer.cs
@@ -1,12 +1,12 @@
-﻿namespace ClrHeapAllocationAnalyzer
-{
-    using System;
-    using System.Collections.Immutable;
-    using Microsoft.CodeAnalysis;
-    using Microsoft.CodeAnalysis.CSharp;
-    using Microsoft.CodeAnalysis.CSharp.Syntax;
-    using Microsoft.CodeAnalysis.Diagnostics;
+﻿using System;
+using System.Collections.Immutable;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Diagnostics;
 
+namespace ClrHeapAllocationAnalyzer.Analyzers
+{
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
     public sealed class ExplicitAllocationAnalyzer : AllocationAnalyzer
     {
