+++ ClrHeapAllocationsAnalyzer/Analyzers/TypeConversionAllocationAnalyzer.cs
@@ -1,14 +1,13 @@
-﻿namespace ClrHeapAllocationAnalyzer
+﻿using System;
+using System.Collections.Immutable;
+using System.Threading;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Diagnostics;
+
+namespace ClrHeapAllocationAnalyzer.Analyzers
 {
-    using System;
-    using System.Collections.Immutable;
-    using System.Threading;
-    using Microsoft.CodeAnalysis;
-    using Microsoft.CodeAnalysis.CSharp;
-    using Microsoft.CodeAnalysis.CSharp.Syntax;
-    using Microsoft.CodeAnalysis.Diagnostics;
-
-
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
     public sealed class TypeConversionAllocationAnalyzer : AllocationAnalyzer
     {
