+++ ConfigurationFileGenerator/Program.cs
@@ -4,6 +4,7 @@ using System.Threading;
 using Buildalyzer;
 using Buildalyzer.Workspaces;
 using ClrHeapAllocationAnalyzer;
+using ClrHeapAllocationAnalyzer.Support;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
